Add unit tests
This makes it easier to make changes in the future, because you have more
confidence that the changes you've made haven't broken anything.

In order to be able to test the "pull words out of this title" unit, I had to
refactor things a bit.  I believe I haven't broken anything while doing so.