Refactor extractWords() to use regular expressions
Using a regex to define a character set you want to filter on is much more
concise, and probably performant, then doing a manual for-in process.  It will
also make it easier if at some point in time you want to support unicode
characters, or other whitespace characters, or things like that.